# CountriesTestApplication/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "REST API Countries Exercise",
    "version": "v1"
  },
  "paths": {
    "/api/countries": {
      "get": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "query",
            "name": "region",
            "type": "string"
          },
          {
            "in": "query",
            "name": "language",
            "type": "string"
          },
          {
            "in": "query",
            "name": "search",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "descending",
            "type": "boolean",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/countries/{code}": {
      "get": {
        "tags": [
          "Country"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/regions": {
      "get": {
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/languages": {
      "get": {
        "tags": [
          "Country"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}